+++ ageofvainilla/src/main/java/config/Configuration.java
@@ -73,12 +73,12 @@ public class Configuration {
 		sprites.put(key, ResourceUtils.getSprite(FETCH(key)));
 	}
 
-	public static double getDisplayWidth() {
-		return getValue("screenWidth");
+	public static int getDisplayWidth() {
+		return getValue("screenWidth").intValue();
 	}
 
-	public static double getDisplayHeight() {
-		return getValue("screenHeight");
+	public static int getDisplayHeight() {
+		return getValue("screenHeight").intValue();
 	}
     
 }
